Adding maven structure, so now it could be done using maven: mvn clean compile exec:exec
