--- src/pdftablesample/Column.java
@@ -1,28 +1,0 @@
-package pdftablesample;
-
-public class Column {
-
-    private String name;
-    private float width;
-
-    public Column(String name, float width) {
-        this.name = name;
-        this.width = width;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public float getWidth() {
-        return width;
-    }
-
-    public void setWidth(float width) {
-        this.width = width;
-    }
-}
